client-test/src/unabomber/client/test/GameEngineTest.java
package unabomber.client.test;

import junit.framework.TestCase;

import org.apache.http.HttpResponse;
import org.apache.http.client.methods.HttpUriRequest;
import org.apache.http.params.HttpParams;

import android.location.Location;

import unabomber.client.GameEngine;
import unabomber.client.UnabomberHttpClient;

public class GameEngineTest extends TestCase {	
	private GameEngine gameEngine;
	private MockHttpClient httpClient;
	private static final String DEVICE_ID = "fake-device-id";

	@Override
	protected void setUp() throws Exception {
		httpClient = new MockHttpClient(null);
		gameEngine = new GameEngine(httpClient, DEVICE_ID);
	}

	public void testSendLocationSendsPostRequest() {
		gameEngine.sendLocation(null);
		
		HttpUriRequest request = httpClient.getRequest();
		assertEquals("POST", request.getMethod());
	}
	
	public void testSendLocationRequestContainsCurrentLocation() {
		Location currentLocation = new Location("GPS");
		gameEngine.sendLocation(currentLocation);
		
		HttpParams params = httpClient.getRequest().getParams();
		assertEquals(currentLocation, (Location)(params.getParameter("current_location")));
	}
	
	public void testSendLocationRequestIsToTheProperUri() {
		gameEngine.sendLocation(null);
		
		HttpUriRequest request = httpClient.getRequest();
		assertEquals("http://localhost:3000/players/" + DEVICE_ID + "/update", request.getURI().toString());
	}
	
	public final class MockHttpClient extends UnabomberHttpClient {
		private HttpResponse response;
		private HttpUriRequest request;
		
		private MockHttpClient(HttpResponse response) {
			this.response = response;
		}
		
		public HttpUriRequest getRequest() {
			return request;
		}

		@Override
		public HttpResponse executeRequest(HttpUriRequest request) {
			this.request = request;
			return response;
		}
	}
}
